Testing division in a box. Main change: made sure that output particle vector is updated newest version before printing ... appears to be working

+++ src/system/Simulation.h
@@ -51,7 +51,7 @@ class Simulation : virtual public Particle{
 
         // Methods for particle dynamics
         void move();
-        void populationDynamics(double);
+        void populationDynamics(int);
 
         // Methods for getting sim data
         int popSize(void){ return particles.size() - boundarysize;}
